Artifice_Nexus-main/backend/requirements.txt
flask
flask-cors
tensorflow
opencv-python
pillow
pyttsx3
huggingface_hub
numpy
gunicorn
